Wp 1312

+++ code/wp-content/plugins/the-events-calendar/src/Tribe/Aggregator/Record/Queue.php
@@ -2,11 +2,14 @@
 // Don't load directly
 defined( 'WPINC' ) or die;
 
-class Tribe__Events__Aggregator__Record__Queue {
+class Tribe__Events__Aggregator__Record__Queue implements Tribe__Events__Aggregator__Record__Queue_Interface {
 	public static $in_progress_key = 'tribe_aggregator_queue_';
 	public static $queue_key = 'queue';
 	public static $activity_key = 'activity';
 
+	/**
+	 * @var Tribe__Events__Aggregator__Record__Abstract
+	 */
 	public $record;
 
 	public $is_fetching = false;
@@ -49,6 +52,11 @@ class Tribe__Events__Aggregator__Record__Queue {
 	 * @var bool
 	 */
 	protected $null_process = false;
+
+	/**
+	 * @var bool Whether this queue instance has acquired the lock or not.
+	 */
+	protected $has_lock = false;
 
 	/**
 	 * Tribe__Events__Aggregator__Record__Queue constructor.
@@ -113,7 +121,7 @@ class Tribe__Events__Aggregator__Record__Queue {
 		}
 	}
 
-	public function init_queue( $items ) {
+	protected function init_queue( $items ) {
 		if ( 'csv' === $this->record->origin ) {
 			$this->record->reset_tracking_options();
 			$this->importer = $items;
@@ -127,8 +135,13 @@ class Tribe__Events__Aggregator__Record__Queue {
 		}
 	}
 
-	public function load_queue() {
-		$this->items = $this->record->meta[ self::$queue_key ];
+	protected function load_queue() {
+		if ( empty( $this->record->meta[ self::$queue_key ] ) ) {
+			$this->is_fetching = false;
+			$this->items       = array();
+		} else {
+			$this->items = $this->record->meta[ self::$queue_key ];
+		}
 
 		if ( 'fetch' === $this->items ) {
 			$this->is_fetching = true;
@@ -155,7 +168,7 @@ class Tribe__Events__Aggregator__Record__Queue {
 	 *
 	 * @return boolean
 	 */
-	public function is_fetching() {
+	protected function is_fetching() {
 		return $this->is_fetching;
 	}
 
@@ -165,16 +178,17 @@ class Tribe__Events__Aggregator__Record__Queue {
 	 * @return int
 	 */
 	public function count() {
-		return count( $this->items );
+		return is_array( $this->items ) ? count( $this->items ) : 0;
 	}
 
 	/**
-	 * Shortcut to check if this queue is empty
+	 * Shortcut to check if this queue is empty.
 	 *
-	 * @return boolean
+	 * @return boolean `true` if this queue instance has acquired the lock and
+	 *                 the count is 0, `false` otherwise.
 	 */
 	public function is_empty() {
-		return 0 === $this->count();
+		return $this->has_lock && 0 === $this->count();
 	}
 
 	/**
@@ -182,7 +196,7 @@ class Tribe__Events__Aggregator__Record__Queue {
 	 *
 	 * @return int
 	 */
-	public function get_total() {
+	protected function get_total() {
 		return $this->count() + $this->activity->count( $this->get_queue_type() );
 	}
 
@@ -191,8 +205,14 @@ class Tribe__Events__Aggregator__Record__Queue {
 	 *
 	 * @return self
 	 */
-	public function save() {
+	protected function save() {
 		$this->record->update_meta( self::$activity_key, $this->activity );
+
+		/** @var Tribe__Meta__Chunker $chunker */
+		$chunker = tribe( 'chunker' );
+		// this data has the potential to be very big, so we register it for possible chunking in the db
+		$key = Tribe__Events__Aggregator__Record__Abstract::$meta_key_prefix . self::$queue_key;
+		$chunker->register_chunking_for( $this->record->post->ID, $key );
 
 		if ( empty( $this->items ) ) {
 			$this->record->delete_meta( self::$queue_key );
@@ -203,7 +223,8 @@ class Tribe__Events__Aggregator__Record__Queue {
 		// If we have a parent also update that
 		if ( ! empty( $this->record->post->post_parent ) ) {
 			$parent = Tribe__Events__Aggregator__Records::instance()->get_by_post_id( $this->record->post->post_parent );
-			if ( isset( $parent->meta[ self::$activity_key ] ) ) {
+
+			if ( ! tribe_is_error( $parent ) && isset( $parent->meta[ self::$activity_key ] ) ) {
 				$activity = $parent->meta[ self::$activity_key ];
 
 				if ( $activity instanceof Tribe__Events__Aggregator__Record__Activity ) {
@@ -224,6 +245,8 @@ class Tribe__Events__Aggregator__Record__Queue {
 
 		wp_update_post( $args );
 
+		$this->release_lock();
+
 		return $this;
 	}
 
@@ -237,44 +260,62 @@ class Tribe__Events__Aggregator__Record__Queue {
 			return $this;
 		}
 
-		if ( $this->is_fetching() ) {
-			$data = $this->record->prep_import_data();
+		$this->has_lock = $this->acquire_lock();
 
-			if (
-				'fetch' === $data
-				|| ! is_array( $data )
-				|| is_wp_error( $data )
-			) {
-				return $this->activity();
+		if ( $this->has_lock ) {
+			if ( $this->is_fetching() ) {
+				if ( $this->record->should_queue_import() ) {
+					$response = $this->record->queue_import();
+
+					if ( $response instanceof WP_Error ) {
+						// the import queueing generated an error
+						$this->record->set_status_as_failed( $response );
+
+						return $this;
+					}
+
+					if ( is_numeric( $response ) ) {
+						// the import queueing was rescheduled
+						$this->record->set_status_as_pending();
+
+						return $this;
+					}
+				}
+
+				$data = $this->record->prep_import_data();
+
+				if (
+					'fetch' === $data
+					|| ! is_array( $data )
+					|| is_wp_error( $data )
+				) {
+					$this->release_lock();
+					return $this->activity();
+				}
+
+				$this->init_queue( $data );
+				$this->save();
 			}
 
-			$this->init_queue( $data );
-			$this->save();
-		}
 
-		// Every time we are about to process we reset the next var
-		$this->next = array();
-
-		if ( ! $batch_size ) {
-			$batch_size = apply_filters( 'tribe_aggregator_batch_size', Tribe__Events__Aggregator__Record__Queue_Processor::$batch_size );
-		}
-
-		for ( $i = 0; $i < $batch_size; $i++ ) {
-			if ( 0 === count( $this->items ) ) {
-				break;
+			if ( ! $batch_size ) {
+				$batch_size = apply_filters( 'tribe_aggregator_batch_size', Tribe__Events__Aggregator__Record__Queue_Processor::$batch_size );
 			}
 
-			// Remove the Event from the Items remaining
-			$this->next[] = array_shift( $this->items );
-		}
+			// Every time we are about to process we reset the next var
+			$this->next = array_splice( $this->items, 0, $batch_size );
 
-		if ( 'csv' === $this->record->origin ) {
-			$activity = $this->record->continue_import();
+			if ( 'csv' === $this->record->origin ) {
+				$activity = $this->record->continue_import();
+			} else {
+				$activity = $this->record->insert_posts( $this->next );
+			}
+
+			$this->activity = $this->activity()->merge( $activity );
 		} else {
-			$activity = $this->record->insert_posts( $this->next );
+			// this queue instance should not register any new activity
+			$this->activity = $this->activity();
 		}
-
-		$this->activity = $this->activity()->merge( $activity );
 
 		return $this->save();
 	}
@@ -304,6 +345,10 @@ class Tribe__Events__Aggregator__Record__Queue {
 	 * execution hangs half way through the processing of a batch.
 	 */
 	public function set_in_progress_flag() {
+		if ( empty( $this->record->id ) ) {
+			return;
+		}
+
 		Tribe__Post_Transient::instance()->set( $this->record->id, self::$in_progress_key, true, HOUR_IN_SECONDS );
 	}
 
@@ -311,6 +356,10 @@ class Tribe__Events__Aggregator__Record__Queue {
 	 * Clears the in progress flag.
 	 */
 	public function clear_in_progress_flag() {
+		if ( empty( $this->record->id ) ) {
+			return;
+		}
+
 		Tribe__Post_Transient::instance()->delete( $this->record->id, self::$in_progress_key );
 	}
 
@@ -320,6 +369,10 @@ class Tribe__Events__Aggregator__Record__Queue {
 	 * @return bool
 	 */
 	public function is_in_progress() {
+		if ( empty( $this->record->id ) ) {
+			return false;
+		}
+
 		Tribe__Post_Transient::instance()->get( $this->record->id, self::$in_progress_key );
 	}
 
@@ -331,11 +384,103 @@ class Tribe__Events__Aggregator__Record__Queue {
 	public function get_queue_type() {
 		$item_type = Tribe__Events__Main::POSTTYPE;
 
-		if ( 'csv' === $this->record->origin ) {
+		if ( ! empty( $this->record->origin ) && 'csv' === $this->record->origin ) {
 			$item_type = $this->record->meta['content_type'];
 		}
 
 		return $item_type;
 	}
-}
 
+	/**
+	 * Acquires the global (db stored) queue lock if available.
+	 *
+	 * @since 4.5.12
+	 *
+	 * @return bool Whether the lock could be acquired or not if another instance/process has
+	 *              already acquired the lock.
+	 */
+	protected function acquire_lock() {
+		if ( empty( $this->record->post->ID ) ) {
+			return false;
+		}
+
+		$post_id = $this->record->post->ID;
+
+		$post_transient = Tribe__Post_Transient::instance();
+
+		$locked = $post_transient->get( $post_id, 'aggregator_queue_lock' );
+
+		if ( ! empty( $locked ) ) {
+			return false;
+		}
+
+		$post_transient->set( $post_id, 'aggregator_queue_lock', '1', 180 );
+
+		return true;
+	}
+
+	/**
+	 * Release the queue lock if this instance of the queue holds it.
+	 *
+	 * @since 4.5.12
+	 *
+	 * @return bool
+	 */
+	protected function release_lock() {
+		if ( empty( $this->record->post->ID ) || ! $this->has_lock ) {
+			return false;
+		}
+
+		$post_id = $this->record->post->ID;
+
+		$post_transient = Tribe__Post_Transient::instance();
+
+		$post_transient->delete( $post_id, 'aggregator_queue_lock' );
+
+		return true;
+	}
+
+	/**
+	 * Whether the current queue process is stuck or not.
+	 *
+	 * @since 4.6.21
+	 *
+	 * @return mixed
+	 */
+	public function is_stuck() {
+		return false;
+	}
+
+	/**
+	 * Orderly closes the queue process.
+	 *
+	 * @since 4.6.21
+	 *
+	 * @return bool
+	 */
+	public function kill_queue() {
+		return true;
+	}
+
+	/**
+	 * Whether the current queue process failed or not.
+	 *
+	 * @since 4.6.21
+	 *
+	 * @return bool
+	 */
+	public function has_errors() {
+		return false;
+	}
+
+	/**
+	 * Returns the queue error message.
+	 *
+	 * @since 4.6.21
+	 *
+	 * @return string
+	 */
+	public function get_error_message() {
+		return '';
+	}
+}
