+++ code/wp-content/plugins/w3-total-cache/Cdn_Plugin.php
@@ -29,38 +29,18 @@ class Cdn_Plugin {
 	 */
 	function run() {
 		$cdn_engine = $this->_config->get_string( 'cdn.engine' );
-		if ( Cdn_Util::is_engine_fsd( $cdn_engine ) ) {
-			$this->run_fsd();
-			return;
-		}
 
 		add_filter( 'cron_schedules', array(
 				$this,
 				'cron_schedules'
 			) );
 
-		if ( !$this->_config->get_boolean( 'cdn.debug' ) )
-			add_filter( 'w3tc_footer_comment', array(
-					$this,
-					'w3tc_footer_comment'
-				) );
+		add_filter( 'w3tc_footer_comment', array(
+				$this,
+				'w3tc_footer_comment'
+			) );
 
 		if ( !Cdn_Util::is_engine_mirror( $cdn_engine ) ) {
-			add_action( 'delete_attachment', array(
-					$this,
-					'delete_attachment'
-				) );
-
-			add_filter( 'update_attached_file', array(
-					$this,
-					'update_attached_file'
-				) );
-
-			add_filter( 'wp_update_attachment_metadata', array(
-					$this,
-					'update_attachment_metadata'
-				) );
-
 			add_action( 'w3_cdn_cron_queue_process', array(
 					$this,
 					'cron_queue_process'
@@ -80,7 +60,17 @@ class Cdn_Plugin {
 					$this,
 					'update_feedback'
 				) );
+
 		}
+
+		add_action( 'delete_attachment',
+			array( $this, 'delete_attachment' ) );
+
+		add_filter( 'update_attached_file',
+			array( $this, 'update_attached_file' ) );
+
+		add_filter( 'wp_update_attachment_metadata',
+			array( $this, 'update_attachment_metadata' ) );
 
 		add_filter( 'w3tc_admin_bar_menu',
 			array( $this, 'w3tc_admin_bar_menu' ) );
@@ -89,6 +79,11 @@ class Cdn_Plugin {
 			add_action( 'w3tc_config_ui_save-w3tc_cdn', array(
 					$this, 'change_canonical_header' ), 0, 0 );
 			add_filter( 'w3tc_module_is_running-cdn', array( $this, 'cdn_is_running' ) );
+		}
+
+		if ( !is_admin() || $this->_config->get_boolean( 'cdn.admin.media_library' ) ) {
+			add_filter( 'wp_prepare_attachment_for_js',
+				array( $this, 'wp_prepare_attachment_for_js' ), 0 );
 		}
 
 		/**
@@ -104,35 +99,6 @@ class Cdn_Plugin {
 					'media_row_actions'
 				), 0, 2 );
 		}
-	}
-
-	/**
-	 * run code for FSD CDN
-	 */
-	private function run_fsd() {
-		add_action( 'w3tc_flush_all', array(
-				'\W3TC\Cdn_Fsd_CacheFlush',
-				'w3tc_flush_all'
-			), 3000, 1 );
-		add_action( 'w3tc_flush_post', array(
-				'\W3TC\Cdn_Fsd_CacheFlush',
-				'w3tc_flush_post'
-			), 3000, 1 );
-		add_action( 'w3tc_flushable_posts', '__return_true', 3000 );
-		add_action( 'w3tc_flush_posts', array(
-				'\W3TC\Cdn_Fsd_CacheFlush',
-				'w3tc_flush_all'
-			), 3000 );
-		add_action( 'w3tc_flush_url', array(
-				'\W3TC\Cdn_Fsd_CacheFlush',
-				'w3tc_flush_url'
-			), 3000, 1 );
-		add_filter( 'w3tc_flush_execute_delayed_operations', array(
-				'\W3TC\Cdn_Fsd_CacheFlush',
-				'w3tc_flush_execute_delayed_operations'
-			), 3000 );
-
-		Util_AttachToActions::flush_posts_on_actions();
 	}
 
 	/**
@@ -187,7 +153,12 @@ class Cdn_Plugin {
 
 		$results = array();
 
-		$common->upload( $files, true, $results );
+		$cdn_engine = $this->_config->get_string( 'cdn.engine' );
+		if ( Cdn_Util::is_engine_mirror( $cdn_engine ) ) {
+			$common->purge( $files, $results );
+		} else {
+			$common->upload( $files, true, $results );
+		}
 
 		return $attached_file;
 	}
@@ -206,7 +177,12 @@ class Cdn_Plugin {
 
 		$results = array();
 
-		$common->delete( $files, true, $results );
+		$cdn_engine = $this->_config->get_string( 'cdn.engine' );
+		if ( Cdn_Util::is_engine_mirror( $cdn_engine ) ) {
+			$common->purge( $files, $results );
+		} else {
+			$common->delete( $files, true, $results );
+		}
 	}
 
 	/**
@@ -224,7 +200,12 @@ class Cdn_Plugin {
 
 		$results = array();
 
-		$common->upload( $files, true, $results );
+		$cdn_engine = $this->_config->get_string( 'cdn.engine' );
+		if ( Cdn_Util::is_engine_mirror( $cdn_engine ) ) {
+			$common->purge( $files, $results );
+		} else {
+			$common->upload( $files, true, $results );
+		}
 
 		return $metadata;
 	}
@@ -685,6 +666,54 @@ class Cdn_Plugin {
 		$admin->change_canonical_header();
 	}
 
+	/**
+	 * Adjusts attachment urls to cdn. This is for those who rely on
+	 * wp_prepare_attachment_for_js()
+	 *
+	 * @param 	array   $response	Mixed collection of data about the attachment object
+	 * @return 	array
+	 */
+	public function wp_prepare_attachment_for_js( $response ) {
+		$response['url'] = $this->wp_prepare_attachment_for_js_url( $response['url'] );
+		$response['link'] = $this->wp_prepare_attachment_for_js_url( $response['link'] );
+
+		if ( !empty( $response['sizes'] ) ) {
+			foreach( $response['sizes'] as $size => &$data ) {
+				$data['url'] = $this->wp_prepare_attachment_for_js_url( $data['url'] );
+			}
+		}
+
+		return $response;
+	}
+
+	/**
+	 * An attachment's local url to modify into a cdn url
+	 *
+	 * @param 	string   $url	the local url to modify
+	 * @return 	string
+	 */
+	private function wp_prepare_attachment_for_js_url( $url ) {
+		$url = trim( $url );
+		if ( !empty( $url ) ) {
+			$parsed = parse_url( $url );
+			$uri = ( isset( $parsed['path'] ) ? $parsed['path'] : '/' ) .
+					   ( isset( $parsed['query'] ) ? '?' . $parsed['query'] : '' );
+
+			$wp_upload_dir = wp_upload_dir();
+			$upload_base_url = $wp_upload_dir['baseurl'];
+
+			if ( substr($url, 0, strlen( $upload_base_url ) ) == $upload_base_url ) {
+				$common = Dispatcher::component( 'Cdn_Core' );
+				$new_url = $common->url_to_cdn_url( $url, $uri );
+				if ( !is_null( $new_url ) ) {
+					$url = $new_url;
+				}
+			}
+		}
+
+		return $url;
+	}
+
 	public function w3tc_admin_bar_menu( $menu_items ) {
 		$cdn_engine = $this->_config->get_string( 'cdn.engine' );
 
@@ -724,14 +753,8 @@ class Cdn_Plugin {
 				sprintf( ' (%s)', $this->cdn_reject_reason ) ) );
 
 		if ( $this->_config->get_boolean( 'cdn.debug' ) ) {
-			$strings[] = "CDN debug info:";
-			$strings[] = sprintf( "%s%s", str_pad( 'Engine: ', 20 ),
-				$this->_config->get_string( 'cdn.engine' ) );
-
-			if ( $this->cdn_reject_reason ) {
-				$strings[] = sprintf( "%s%s", str_pad( 'Reject reason: ', 20 ),
-					$this->cdn_reject_reason );
-			}
+			$strings[] = '';
+			$strings[] = 'CDN debug info:';
 
 			if ( count( $this->_replaced_urls ) ) {
 				$strings[] = "Replaced URLs:";
@@ -742,6 +765,7 @@ class Cdn_Plugin {
 						Util_Content::escape_comment( $new_url ) );
 				}
 			}
+			$strings[] = '';
 		}
 
 		return $strings;
@@ -1130,15 +1154,8 @@ class _Cdn_Plugin_ContentFilter {
 	 */
 	function _link_replace_callback_ask_cdn( $match, $quote, $url, $path ) {
 		$common = Dispatcher::component( 'Cdn_Core' );
-		$cdn = $common->get_cdn();
-		$remote_path = $common->uri_to_cdn_uri( $path );
-		$new_url = $cdn->format_url( $remote_path );
-		if ( $new_url ) {
-			$is_engine_mirror = Cdn_Util::is_engine_mirror(
-				$this->_config->get_string( 'cdn.engine' ) );
-
-			$new_url = apply_filters( 'w3tc_cdn_url', $new_url, $url,
-				$is_engine_mirror );
+		$new_url = $common->url_to_cdn_url( $url, $path );
+		if ( !is_null( $new_url ) ) {
 			$this->replaced_urls[$url] = $new_url;
 			return $quote . $new_url;
 		}
